+++ algorithm/sort/includes/SelectionSort.hpp
@@ -1,9 +1,9 @@
 #ifndef SELECTIONSORT_HPP
 #define SELECTIONSORT_HPP
 #include <vector>
-#include <cstdint>
+#include <cstddef>
 
-void SelectionSort(std::vector <uint64_t> & arr);
+void SelectionSort(std::vector <size_t> & arr);
 
 #endif /* SELECTIONSORT_HPP */
